pyproject.toml
[project]
name = "perplex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.6.0",
    "beautifulsoup4>=4.13.3",
    "dotenv>=0.9.9",
    "duckduckgo-search>=7.5.5",
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "groq>=0.20.0",
    "llama-index>=0.12.27",
    "llama-index-core>=0.12.27",
    "llama-index-llms-groq>=0.3.1",
    "llama-index-llms-openai>=0.3.29",
    "llama-index-tools-tavily-research>=0.3.0",
    "lxml-html-clean>=0.4.1",
    "readability-lxml>=0.8.1",
    "requests>=2.32.3",
    "requests-cache>=1.2.1",
    "sentence-transformers>=4.0.1",
    "soundfile>=0.13.1",
    "streamlit>=1.44.0",
    "streamlit-audiorec>=0.1.3",
    "torchaudio>=2.6.0",
    "uvicorn>=0.34.0",
]
